pyproject.toml
[tool.poetry]
name = "bot-example"
version = "0.1.0"
description = ""
authors = ["Sergio Soto <sergiosotonunez@gmail.com>"]
packages = [
    { include = "mybot" },
]


[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
python-telegram-bot = "^13.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
